chore: refactor user_operator endpoint
