+++ bc_obps/registration/api/__init__.py
@@ -8,4 +8,18 @@ from . import (
     regulated_products,
     reporting_activities,
 )
+
+from .user_operator import (
+    operator,
+    is_approved_admin_user_operator,
+    operator_access_declined,
+    operator_has_admin,
+    user_operator_from_user,
+    user_operator_id,
+    user_operator_initial_requests,
+    user_operator_list_from_user,
+    user_operator_operator,
+)
+from .select_operator import request_access, request_admin_access
+from .select_operator.user_operator import update_status, user_operator_id
 from .api_base import router  # django_ninja router object
