+++ bc_obps/registration/api/user_operator/operator_has_admin.py
@@ -1,0 +1,25 @@
+from uuid import UUID
+from registration.service.user_operator.GetUserOperatorAdminExistsService import GetUserOperatorAdminExistsService
+
+from registration.decorators import authorize, try_except_wrapper
+from registration.schema import (
+    Message,
+)
+from registration.api.api_base import router
+
+from registration.models import (
+    AppRole,
+    UserOperator,
+)
+from ninja.responses import codes_4xx
+
+
+@router.get(
+    "/operator-has-admin/{operator_id}",
+    response={200: bool, codes_4xx: Message},
+    url_name="get_user_operator_admin_exists",
+)
+@authorize(AppRole.get_all_authorized_app_roles(), UserOperator.get_all_industry_user_operator_roles(), False)
+@try_except_wrapper()
+def get_user_operator_admin_exists(request, operator_id: UUID):
+    return GetUserOperatorAdminExistsService.get_user_operator_admin_exists(request, operator_id)
