+++ bc_obps/registration/service/user_operator/GetUserOperatorAccessDeclinedService.py
@@ -1,0 +1,15 @@
+from uuid import UUID
+
+from registration.models import (
+    User,
+    UserOperator,
+)
+
+
+class GetUserOperatorAccessDeclinedService:
+    def get_user_operator_access_declined(request, operator_id: UUID):
+        user: User = request.current_user
+        is_declined = UserOperator.objects.filter(
+            operator_id=operator_id, user_id=user.user_guid, status=UserOperator.Statuses.DECLINED
+        ).exists()
+        return 200, is_declined
