+++ bc_obps/registration/service/user_operator/UserOperatorFromUserService.py
@@ -1,0 +1,21 @@
+from registration.schema import (
+    PendingUserOperatorOut,
+)
+
+from registration.models import (
+    UserOperator,
+)
+
+
+class UserOperatorFromUserService:
+    def get_user_operator_from_user(request):
+        try:
+            user_operator = (
+                UserOperator.objects.only("id", "status", "operator__id", "operator__is_new", "operator__status")
+                .exclude(status=UserOperator.Statuses.DECLINED)
+                .select_related("operator")
+                .get(user_id=request.current_user.user_guid)
+            )
+        except UserOperator.DoesNotExist:
+            return 404, {"message": "User is not associated with any operator"}
+        return 200, PendingUserOperatorOut.from_orm(user_operator)
